+++ Public/Journey/New-MermaidJourney.ps1
@@ -1,0 +1,23 @@
+function New-MermaidJourney {
+    [CmdletBinding()]
+    param (
+        $Title,
+        [string[]] $Section
+    )
+    
+    begin {
+        $output = [System.Text.StringBuilder]::new()
+        $output.AppendLine("journey")  | Out-Null
+    }
+    
+    process {
+        $output.AppendLine("`ttitle $Title")  | Out-Null
+        foreach($SectionElement in $Section){
+            $output.AppendLine($SectionElement) |Out-Null
+        }
+    }
+    
+    end {
+        return $output.ToString()
+    }
+}